requirements.txt
cliff>=1.10.0  # Apache-2.0
pbr>=0.11,<2.0
os-client-config
python-openstacksdk
networkx
paramiko
python-heatclient
AttrDict
appdirs
pyaml
